% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VT_Color_Palette.R
\name{scale_color_vt}
\alias{scale_color_vt}
\title{Color scale constructor for vt colors}
\usage{
scale_color_vt(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in vt_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradient(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for vt colors
}
