% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VT_Color_Palette.R
\name{vt_cols}
\alias{vt_cols}
\title{Function to extract VT colors as hex codes}
\usage{
vt_cols(...)
}
\arguments{
\item{...}{Character names of vt_colors}
}
\description{
Function to extract VT colors as hex codes
}
