% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VT_Color_Palette.R
\name{vt_pal}
\alias{vt_pal}
\title{Return function to interpolate a VT color palette}
\usage{
vt_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in vt_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a VT color palette
}
